+++ My_first_Cpp_game/renderer.h
@@ -14,6 +14,6 @@ void drawTable(int mode,
 				u32* cell_color, 
 				const u32 number_of_cell_color, 
 				u32 background_color, 
-				int** scores, 
+				u64** scores, 
 				u32 number_color);
 void showCursorPosition();
